trade Journal 2.0/tradezella-clone/mac/requirements.txt
# Mac-side dependencies (receiver script + sync)
# Install with: pip install -r requirements.txt

httpx==0.28.1         # async-capable HTTP client for calling the VPS
python-dotenv==1.0.1  # loads .env file into environment variables

# For running tests
pytest==8.3.4
pytest-asyncio==0.24.0
respx==0.22.0         # mock library for httpx requests in tests
